bife part 2 :D

+++ src/Tool_Ranged.h
@@ -8,7 +8,7 @@
 class Tool_Ranged : public Tool{
     float ProjectileSize, ProjectileSpeed, Lifetime;
 public:
-    Tool_Ranged(const std::string& name_, float Damage_, float Cooldown_, float Range_ = 1.0f, float Critical_Chance_ = 0,
+    Tool_Ranged(const std::string& name_, float Damage_, float Cooldown_, float Range_ = 1.0f, int Critical_Chance_ = 0,
     float ProjectileSize_ = 30.0f, float ProjectileSpeed_ = 15.0f, float Lifetime_ = 4.0f);
 
     void CreateAttackHitbox(sf::Vector2f Position, sf::Vector2f Offset, sf::Angle Degrees) override;
